chore: split module for transaction submission command
